Crate dh_drone_stm32f4_utils

A crate of common code for STM32F4xx microcontrollers using Drone OS.

Usage

Add the crate to your Cargo.toml dependencies:

[dependencies]
drone-stm32f4-utils = { git = "https://github.com/dhylands/drone-stm32f4-utils" }

Modules

clock
Functions for configuring the various clocks used on the STM32F4xx.

consts
Register constants used by this crate.

gpiosignal
Some concrete implementations of the Signal trait for GPIO pins.

signal
A trait for controlling a signal (i.e. a gpio output like a relay or LED).