Trait dh_drone_stm32f4_utils::clock::SystemClock

pub trait SystemClock {
    fn flash_latency(&self) -> u32;

    fn hse_freq(&self) -> u32;

    fn pllm(&self) -> u32;

    fn plln(&self) -> u32;

    fn pllp(&self) -> u32;

    fn pllq(&self) -> u32;
}

A trait that provides the constants needed to setup the main clocks used on the STM32F4xx.

These clocks are:

System clock. This is used to derive the AHB, APB1, and APB2 clocks.

Bus clock (also referred to as AHB or HCLK) which is the clock that the CPU runs at.

Low speed peripheral clock (APB1).

High speed peripheral clock (APB2).

USB/SDIO/RNG clock.

The VCO-clock is calculated as the HSE-frequency * PLLN / PLLM.

The system-clock is calculated as the VCO-clock / PLLP.

The USB/SDIO/RNG-clock is calclated as the system-clock / PLLQ.

The AHB clock is equal to the system clock.

The APB1 clock is equal to the AHB clock divided by 4.

The APB2 clock is equal to the AHB clock divided by 2.

There are lots of constraints on the frequencies that the various clocks can have so you should consult the datasheet for the particular processor you're using to determine these constraints.

Example

This example configures the WeAct STM32F411CEU6 board to run at 96 MHz. The WeAct STM32F411CEU6 board has a 25 MHz HSE crystal.

impl SystemClock for WeActSystemClock {
    fn flash_latency(&self) -> u32 {
        3
    }

    fn hse_freq(&self) -> u32 {
        25_000_000
    }

    fn pllm(&self) -> u32 {
        25
    }

    fn plln(&self) -> u32 {
        192
    }

    fn pllp(&self) -> u32 {
        2
    }

    fn pllq(&self) -> u32 {
        4
    }
}

This will result in the following frequencies:

Clock	Frequency

SYSCLK	96 MHz

AHB	96 MHz

APB1	24 MHz

APB2	48 MHz

USB	48 MHz

While the STM32F411 can go upto 100 MHz, we limit it to 96 MHz so that we can get a "nice" frequency for the USB clock (which should be a multiple of 24 MHz).

Required methods

fn flash_latency(&self) -> u32

Specifies the number of wait states to set for FLASH_ACR register. This will depend on the MCU voltage and system clock speed.

fn hse_freq(&self) -> u32

The frequency of the external high speed crystal.

fn pllm(&self) -> u32

PLLM - Division factor for the main PLL (PLL) input clock.

fn plln(&self) -> u32

PLLN - PLL multiplication factor for VCO. The VCO frequency is calculates as HSE frequency * PLLN / PLLM.

fn pllp(&self) -> u32

PLLP - PLL division factor for main system clock.

fn pllq(&self) -> u32

PLLQ - PLL division factor for USB OTG FS, and SDIO clocks

Loading content...