Struct dh_drone_stm32f4_utils::gpiosignal::GpioSignalActiveHigh

pub struct GpioSignalActiveHigh<PIN> where
    PIN: GpioPinMap,  { /* fields omitted */ }

Implements a Signal which is on when a GPIO pin is at a logic high.

Example

If an LED is wired between CPU pin PB5 and GND, then you would turn the LED on by setting pin PB5 to a logic high.

    let led = GpioSignalActiveHigh::init(periph_gpio_b5!(reg));
    led.on()

Implementations

impl<PIN: GpioPinMap> GpioSignalActiveHigh<PIN>

pub fn init(pin: GpioPinPeriph<PIN>) -> Self

Associate a pin with this Signal. Turning this type of Signal on will set the GPIO Pin high.