Trait dh_drone_stm32f4_utils::signal::Signal

pub trait Signal {
    fn on(&self);

    fn off(&self);
}

A Signal has the notion of being on or off which is independent of how that signal is manifested in the hardware.

Some signals will be on when the output is a logic high and some signals will be on when the output is a logic low.

Required methods

fn on(&self)

Turns the signal on.

fn off(&self)

Turns the signal off.

Loading content...